pyproject.toml
[tool.poetry]
name = "solvemate"
version = "0.1.0"
description = ""
authors = ["myoungjinGo-FE <cokemania2099@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.4"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
mypy = "^1.14.0"
drf-yasg = "^1.21.8"
requests = "^2.32.3"
django-cors-headers = "^4.6.0"
python-dotenv = "^1.0.1"
gunicorn = "^23.0.0"
freezegun = "^1.5.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-django = "^4.9.0"
pytest-cov = "^6.0.0"
types-requests = "^2.32.0.20241016"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
